requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
asyncio>=3.4.3
flask>=2.0.0
flask-cors>=3.0.10
flask-limiter>=3.3.0
markdown2>=2.4.0
colorlog>=6.7.0
psutil>=5.9.0
sphinx>=4.5.0
sphinx_rtd_theme>=1.0.0
pylint>=2.15.0
coverage>=6.5.0
build>=0.10.0
python-json-logger>=2.0.7